requirements.txt
streamlit
pandas
numpy
plotly
python-dateutil
pyarrow
fastparquet
matplotlib